Comment up the tech demo
